pr #160. Fix missing coverage.
