+++ src/fparser/two/tests/fortran2003/test_format_specification_r1002.py
@@ -279,45 +279,3 @@ def test_syntaxerror_C1002(f2003_create):
         with pytest.raises(NoMatchError):
             _ = Format_Specification(my_input)
 
-
-def test_internal_errors1(f2003_create, monkeypatch):
-    '''Check that an internal error is raised if the length of the Items
-    list is not 2 as the str() method assumes that it is.
-
-    '''
-    line = "2P F2.2"
-    ast = Format_Item_C1002(line)
-    monkeypatch.setattr(ast, "items", [None, None, None])
-    with pytest.raises(InternalError) as excinfo:
-        str(ast)
-    assert "should be of size 2 but found '3'" in str(excinfo)
-
-
-def test_internal_error2(f2003_create, monkeypatch):
-    '''Check that an internal error is raised if the module name (entry 0
-    of items) is empty or None as the str() method assumes that it has
-    content.
-
-    '''
-    line = "2P F2.2"
-    ast = Format_Item_C1002(line)
-    monkeypatch.setattr(ast, "items", [None, ast.items[1]])
-    with pytest.raises(InternalError) as excinfo:
-        str(ast)
-    assert ("items entry 0 should contain a format items object but it "
-            "is empty or None") in str(excinfo)
-
-
-def test_internal_error3(f2003_create, monkeypatch):
-    '''Check that an internal error is raised if the module name (entry 1
-    of items) is empty or None as the str() method assumes that it has
-    content.
-
-    '''
-    line = "2P F2.2"
-    ast = Format_Item_C1002(line)
-    monkeypatch.setattr(ast, "items", [ast.items[0], None])
-    with pytest.raises(InternalError) as excinfo:
-        str(ast)
-    assert ("items entry 1 should contain a format items object but it "
-            "is empty or None") in str(excinfo)
